[develop] - fixed controllers
