+++ src/Portal/Controllers/v1/ZoneController.cs
@@ -10,7 +10,7 @@ using Portal.Services.ZoneService;
 using Portal.Services.ZoneService.Exceptions;
 using Swashbuckle.AspNetCore.Annotations;
 
-namespace Portal.Controllers;
+namespace Portal.Controllers.v1;
 
 /// <summary>
 /// Контроллер зон.
